Refactor serial wrapper as internal package
- Move serial into internal package
- Add comments
- Do not export 'retrieve' function
- Transform magic numbers into readable constants
